Add merge operations to blend attrs provided by selector str and args
Attributes can be passed in as part of the selector string in
the first argument and as an obj in the second argument.
This code merges all those attributes, giving `class` and `style`
attrs special attention. Obj attrs in the second argument are given
priority over seletor attrs in the first arg.